pyproject.toml
[tool.poetry]
name = "almaz-test"
version = "0.1.0"
description = ""
authors = ["Alex Pavlov <yukiu217@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.5"
uvicorn = "^0.32.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.36"}
asyncpg = "^0.30.0"
pydantic-settings = "^2.6.1"
httpx = "^0.27.2"
loguru = "^0.7.2"
alembic = "^1.14.0"
pytest = "^8.3.3"
redis = {extras = ["redis"], version = "^5.2.1"}
fastapi-cache2 = "^0.2.2"
python-multipart = "^0.0.20"
sqlalchemy-utils = "^0.41.2"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
pytest-asyncio = "^0.24.0"
asgi-lifespan = "^2.1.0"
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
